🧪 Add tests for check_factor in factor_theorem.py
- Moved `evaluate_polynomial` and `check_factor` tests from `test_Algebra.py` to `test_factor_theorem.py`.
- Added edge case tests for floating-point imprecision, empty arrays, all zeros, and gaps in powers to `check_factor`.
- Updated `check_factor` to use `math.isclose` for handling floating-point errors.
